maia-sdr-devel: move image to ghcr.io
This moves the maia-sdr-devel image from Docker hub to ghcr.io.
This change is mainly motivated by latest policy changes in Docker
hub regarding organizations.

The image is now built and pushed to GHCR using a Github action.

To build the image, a tag maia-sdr-devel-nnnn shall be pushed.
This will tag the docker image as maia-sdr-devel:nnnn. Alternatively,
for testing it is possible to push to the branch build-maia-sdr-devel.
This will build and push an image tagged as
maia-sdr-devel:test-maia-sdr-devel.

Signed-off-by: Daniel Estévez <daniel@destevez.net>